src/main/webapp/app/entities/graded-comment/delete/graded-comment-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { IGradedComment } from '../graded-comment.model';
import { GradedCommentService } from '../service/graded-comment.service';
import { FaIconComponent } from '@fortawesome/angular-fontawesome';
import { AlertErrorComponent } from '../../../shared/alert/alert-error.component';
import { TranslateDirective } from '../../../shared/language/translate.directive';
import { FormsModule } from '@angular/forms';
import { NgIf } from '@angular/common';

@Component({
  templateUrl: './graded-comment-delete-dialog.component.html',
  standalone: true,
  imports: [NgIf, FormsModule, TranslateDirective, AlertErrorComponent, FaIconComponent],
})
export class GradedCommentDeleteDialogComponent {
  gradedComment?: IGradedComment;

  constructor(
    protected gradedCommentService: GradedCommentService,
    protected activeModal: NgbActiveModal,
  ) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.gradedCommentService.delete(id).subscribe(() => {
      this.activeModal.close('deleted');
    });
  }
}
